feat(control-panel): support deploying large station WASM
