+++ core/control-panel/api/spec.did
@@ -245,12 +245,23 @@ type CanDeployStationResult = variant {
   Err : ApiError;
 };
 
+type WasmModuleExtraChunks = record {
+  // The asset canister from which the chunks are to be retrieved.
+  store_canister : principal;
+  // The list of chunk hashes in the order they should be appended to the wasm module.
+  chunk_hashes_list : vec blob;
+  // The hash of the assembled wasm module.
+  wasm_module_hash : blob;
+};
+
 // The canister modules required for the control panel.
 type UploadCanisterModulesInput = record {
   // The upgrader wasm module to use for the station canister.
   upgrader_wasm_module : opt blob;
   // The station wasm module to use.
   station_wasm_module : opt blob;
+  // Optional extra chunks of the station canister wasm module.
+  station_wasm_module_extra_chunks : opt opt WasmModuleExtraChunks;
 };
 
 // The result of uploading canister modules.
