+++ core/control-panel/impl/src/services/deploy.rs
@@ -11,6 +11,7 @@ use ic_cdk::api::id as self_canister_id;
 use ic_cdk::api::management_canister::main::{self as mgmt};
 use lazy_static::lazy_static;
 use orbit_essentials::api::ServiceResult;
+use orbit_essentials::install_chunked_code::install_chunked_code;
 use std::sync::Arc;
 
 lazy_static! {
@@ -47,8 +48,9 @@ impl DeployService {
         let config = canister_config().ok_or(DeployError::Failed {
             reason: "Canister config not initialized.".to_string(),
         })?;
-        let station_wasm_module = config.station_wasm_module;
         let upgrader_wasm_module = config.upgrader_wasm_module;
+        let station_wasm_module = config.station_wasm_module;
+        let station_wasm_module_extra_chunks = config.station_wasm_module_extra_chunks;
 
         let can_deploy_station_response = user.can_deploy_station();
         match can_deploy_station_response {
@@ -97,11 +99,8 @@ impl DeployService {
             .collect::<Vec<_>>();
 
         // installs the station canister with the associated upgrader wasm module
-        mgmt::install_code(mgmt::InstallCodeArgument {
-            mode: mgmt::CanisterInstallMode::Install,
-            canister_id: station_canister.canister_id,
-            wasm_module: station_wasm_module,
-            arg: Encode!(&station_api::SystemInstall::Init(station_api::SystemInit {
+        let station_install_arg =
+            Encode!(&station_api::SystemInstall::Init(station_api::SystemInit {
                 name: input.name.clone(),
                 admins,
                 upgrader: station_api::SystemUpgraderInput::WasmModule(upgrader_wasm_module),
@@ -111,12 +110,16 @@ impl DeployService {
             }))
             .map_err(|err| DeployError::Failed {
                 reason: err.to_string(),
-            })?,
-        })
+            })?;
+        install_chunked_code(
+            station_canister.canister_id,
+            mgmt::CanisterInstallMode::Install,
+            station_wasm_module,
+            station_wasm_module_extra_chunks,
+            station_install_arg,
+        )
         .await
-        .map_err(|(_, err)| DeployError::Failed {
-            reason: err.to_string(),
-        })?;
+        .map_err(|err| DeployError::Failed { reason: err })?;
 
         self.user_service
             .add_deployed_station(&user.id, station_canister.canister_id, ctx)
